rename reference module to change_reference
