+++ crates/gitbutler-repo/tests/mod.rs
@@ -1,2 +1,2 @@
+mod change_reference;
 mod credentials;
-mod reference;
